+++ src/Triangle.java
@@ -143,5 +143,4 @@ public interface Triangle extends TriangleKernel {
      * @return new {@code Triangle} to represent {@code this} in 2 dimensions.
      */
     double[] twoDimensionRepresentation();
-
 }